Refactor audio handling to BlocklyInterface
Moving implementation for audio handling to BlocklyInterface.
